+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 
 ### Initial Release
 
+## 0.0.1-beta.12 (2020-09-02)
+* Improved robustness if a user did not set a doublePressTime
+* Disabled remote control triplePress for now
+
+## 0.0.1-beta.11 (2020-09-01)
+* Improved generation of UUID, now based on ip address, and not on device name. You will have to remove and re-add your device
+
 ## 0.0.1-beta.11 (2020-09-01)
 * Fixed bug setting accesory icon for Home Theater
 * Fixed some more config.schema typos